Commenting, registration, participation

+++ base/views.py
@@ -1,9 +1,13 @@
 from bdb import GENERATOR_AND_COROUTINE_FLAGS
+from ctypes import sizeof
+from itertools import count
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from unicodedata import name
 from django.shortcuts import render, redirect
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseBadRequest
 from .models import Message, Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
@@ -12,16 +16,19 @@ from django.contrib.auth.models import User
 # Create your views here.
 from django.http import HttpResponse
 
-rooms = [
+""" rooms = [
     {'id': 0, 'name': 'This room does not exist'},
     {'id':1, 'name': 'Common room'},
     {'id': 2, 'name': 'Meeting Room 1'},
     {'id': 3, 'name': 'Meeting Room 2'},
-]
+] """
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username = username)
@@ -34,8 +41,21 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request, "Username or password incorrect")
-    context = {}
+    context = {'page' : page}
     return render(request, 'base/login_register.html', context)
+
+def registerUser(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit = False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+        else: 
+            messages.error(request, 'Registration unsuccessful')
+    return render(request, 'base/login_register.html', {'form' : form})
 
 def logoutUser(request):
     logout(request)
@@ -51,11 +71,21 @@ def home(request):
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
-    context = {'rooms' : Room.objects.get(id = pk) if len(Room.objects.filter(id = pk)) > 0 else "Room does not exist", 'messages' : list(Message.objects.all())}
+    room = Room.objects.get(id = pk)
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room, 
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    context = {'room' : room, 'usermessages' : room.message_set.all().order_by('-created'), 'participants' : room.participants.all()}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url = '/login')
 def createRoom(request):
-    form = RoomForm
+    form = RoomForm({'host' : request.user})
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -64,9 +94,12 @@ def createRoom(request):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url= '/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+    if request.user != room.host:
+        return HttpResponse('Request not valid')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
         if form.is_valid():
@@ -75,9 +108,22 @@ def updateRoom(request, pk):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = '/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id = pk)
+    if request.user != room.host:
+        return HttpResponse('Request not valid')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj' : room})
+
+@login_required(login_url = '/login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id = pk)
+    if request.user != message.user:
+        return HttpResponse('Not your comment to delete')
+    if request.method == 'POST':
+        message.delete()
+        return redirect('room', pk = message.room.id)
+    return render(request, 'base/delete.html', {'obj' : message})
